Consolidate object models to one file
