+++ examples/react-native/v12/TestApp/src/components/subscribe-api/AlwaysWait.tsx
@@ -4,7 +4,7 @@ import {WaitForSync} from 'realm';
 import {useRealm, useQuery} from '@realm/react';
 import {View, Text, FlatList} from 'react-native';
 
-import {Bird} from '../../models/Bird';
+import {Bird} from '../../models';
 import {Subscription} from 'realm/dist/bundle';
 // :snippet-end:
