+++ examples/react-native/v12/TestApp/src/components/subscribe-api/BasicSubscription.tsx
@@ -3,7 +3,7 @@ import React, {useEffect, useState} from 'react';
 import {useRealm, useQuery} from '@realm/react';
 import {View, Text, FlatList} from 'react-native';
 
-import {Bird} from '../../models/Bird';
+import {Bird} from '../../models';
 import {Subscription} from 'realm/dist/bundle';
 
 export const BasicSubscription = () => {
