--- examples/react-native/v12/TestApp/src/models/Bird.ts
@@ -1,17 +1,0 @@
-import Realm, {BSON, ObjectSchema} from 'realm';
-
-export class Bird extends Realm.Object<Bird> {
-  _id!: BSON.ObjectId;
-  name!: string;
-  haveSeen!: boolean;
-
-  static schema: ObjectSchema = {
-    name: 'Bird',
-    properties: {
-      _id: 'objectId',
-      name: 'string',
-      haveSeen: 'bool',
-    },
-    primaryKey: '_id',
-  };
-}
