--- examples/react-native/v12/TestApp/src/models/Person.ts
@@ -1,17 +1,0 @@
-import Realm, {BSON, ObjectSchema} from 'realm';
-
-export class Person extends Realm.Object {
-  _id!: BSON.ObjectId;
-  name!: string;
-  age!: number;
-
-  static schema: ObjectSchema = {
-    name: 'Person',
-    properties: {
-      _id: 'objectId',
-      name: 'string',
-      age: 'int',
-    },
-    primaryKey: '_id',
-  };
-}
